Progressing on bringing Radiant 1.6.x to a workable state on Windows. This commit: 1. Reverts to using GtkR-deps-1.6-3.zip (instead of GtkR-deps-1.6-4.zip). Some font-related DLLs are therefore no longer used with this patch. 2. Because of the above #1, now using the really old OpenGL font rendering code on Windows systems.  This is controlled via "#ifdef _WIN32" blocks. Linux still uses the new and improved OpenGL font rendering. Note that this old font rendering that Windows now uses (err, has used in the past too) makes use of gdk_gl_font_use_pango_font(), which is really old and crufty.
So, with this commit Radiant is fully working (modulo bugs) on Linux and
Windows, and OpenGL fonts work too.

I WOULD NOT SUGGEST MERGING THIS CHANGE INTO TRUNK.  THE CORRECT APPROACH
IS TO INSTEAD FIX THE DLL LIBRARY DEPENDENCIES IN TRUNK.  I AM ONLY COMMITTING
THIS SO THAT I CAN CONTINUE FIXING BUGS IN RADIANT IN THE MEANTIME.

There are still a zillion bugs on Windows.  For example, I can't even save
a .map file at this point.  That is my next task, to fix saving of .map files.


git-svn-id: svn://svn.icculus.org/gtkradiant/GtkRadiant/branches/Rambetter-temp-fixes@350 8a3a26a2-13c4-0310-b231-cf6edde360e5