Making more progress on	the EXPERIMENTAL_HIGH_PRECISION_MATH_Q3MAP2_FIXES code.
- Renaming CopyWindingAccuToNormal() to	CopyWindingAccuToRegular() in order to
avoid confusing	"normal" (as in	"regular") with	"normal" (as in	the normal
to the plane).

- In mathlib.h/mathlib.c, breaking out all the *Accu() #define functions
into actual functions in mathlib.c.  I think this is much safer	for ensuring
that the correct types are passed in.  Otherwise math might take place
in the 32 bit world without any	warning.

- Adding VectorCopyAccuToRegular() and VectorCopyRegularToAccu() to mathlib.

- Improving comments on	VEC_SMALLEST_EPSILON_AROUND_ONE	and
VEC_ACCU_SMALLEST_EPSILON_AROUND_ONE.

- Had to "adjust" some code in polylib.c due to	the change in functions
in mathlib (moving from	#define	to actual functions).

In case	you're curious as to the correctness of	this code at this point:
I was able to compile Icy Fantasy -bsp stage in	the same amount	of time	that
it took	in the original	q3map2 (23 seconds to be exact), and there are no
more disappearing triangles.  I'm almost positive the regression tests are
fixed at this point, but there is no hurry in testing them because I still
have to	comb through quite a bit of code to ensure that	things are being
done correctly with respect to math accuracy.


git-svn-id: svn://svn.icculus.org/gtkradiant/GtkRadiant/branches/Rambetter-math-fix-experiments@394 8a3a26a2-13c4-0310-b231-cf6edde360e5