Making more progress on EXPERIMENTAL_HIGH_PRECISION_MATH_Q3MAP2_FIXES code.
- Setting EXPERIMENTAL_HIGH_PRECISION_MATH_Q3MAP2_FIXES to 1 (enabled).  I'll
turn this off before merging into trunk.

- In function MapPlaneFromPoints() in map.c, doing a bit tighter math when
EXPERIMENTAL_HIGH_PRECISION_MATH_Q3MAP2_FIXES is on.  This comes at the
expense of a few extra CPU cycles.  The signature of the function remains
the same.

- In VectorNormalize() in mathlib.c, fixing a cast of the return value.
I did not see compile warnings or runtime errors, but it seems wrong the
way it was.

- Adding VectorNormalizeAccu() in mathlib.c, for normalizing vectors built on
the high resolution data type.


git-svn-id: svn://svn.icculus.org/gtkradiant/GtkRadiant/branches/Rambetter-math-fix-experiments@395 8a3a26a2-13c4-0310-b231-cf6edde360e5