Getting ready to commit Rambetter-math-fix-experiments branch changes to trunk, as pertaining to q3map2.  Some last-minute changes:
- EXPERIMENTAL_HIGH_PRECISION_MATH_Q3MAP2_FIXES, EXPERIMENTAL_SNAP_NORMAL_FIX,
and EXPERIMENTAL_SNAP_PLANE_FIX all set to 0 to make sure stuff in trunk 
uses old "broken but stable" code, at least initially until there is further
testing done and/or some decision is made about which code to use.

- Fixing small typo in comment in PlaneEqual() in map.c.

- In comment to BaseWindingForPlane() function, describing why the old
function is broken as far as accuracy goes.  This used to be a TODO.

- VectorIsOnAxialPlane() function comment fixed	in mathlib.c.


git-svn-id: svn://svn.icculus.org/gtkradiant/GtkRadiant/branches/Rambetter-math-fix-experiments@415 8a3a26a2-13c4-0310-b231-cf6edde360e5